[DDW-990] Separate apps for rust and haskell nodes via env flag
We need to target two main versions of the cardano explorer: one for the Rust node (later including byron features) and one for the Haskell node. Both have a slightly (or major) different feature set and maybe also different apis.

There are two ways we could approach the solution: 1. having a single app that gets configured based on an environment flag or 2. having two separate apps that can share some parts.

I have chosen the 2. approach as it makes it easier to reason about each app. The default target is the haskell client (no config needed) - if you pass IS_RUST_CLIENT=1 to any of the yarn commands for dev or build you will get the version targeting the rust client :)

Of course this is just a start and we need to think about shared / dedicated features etc. in the details of the two apps …