[DDW-1020] Implement search block by number logic
This also turned into a major refactoring and cleanup
of the search specs to be more readable and contain
less technical details that are not relevant for
the api integration.

GraphQL fragments have been introduced to reduce
the repeating boilerplate in queries and to improve
separation of concerns. E.g there is a fragment
called BlockDetails which is the full blown representation
of a block fetched from the API. This is used
as a result type for searches by id and number.