[DDW-1020] Cleanup search feature
These changes reduce the number of files in the
search feature to make it more approachable.

The static (global) default setup of the feature
has been replaced with a generator function that
sets up and returns a fully encapsulated instance
of the feature logic which can be reused within
any number of components or instantiated multiple
times with different configurations.