[DDW-1014] Trigger correct query based on search input
The problem was that we have no easy way to
distinguish between transaction and block IDs
and for number inputs we have to ask the user
what he meant at least when there is an epoch
AND block with that number.

This commit solves the tx vs. block by fetching
both and checking the results + redirecting to
the corresponding search result page. The epoch
vs. block number problem was solved in the same
way but not perfectly: if you enter a number where
an epoch AND block exists it currently prefers the
epoch although we need to ask the user in the end.