[DDW-1014] Fix storybook and search perf issues
1. The problem with storybook was that the feature
context was created within the index file where
also the api is imported, which resulted in the
graphql files being imported too, which we haven’t
setup in storybook (and shouldn’t). I separated
out the context and feature hooks into 
[feature]/context.ts files

2. The search performance issues where caused by
(1) not having global state that survives page
redirects. This was solved by moving the global
features to the special _app.tsx (the root for
next.js apps) which is not unmounted on redirects.
(2) the search store now checks if the requested
results are already loaded instead of blindly 
executing the search again.