Ddw 1014 trigger appropriate query based on input analysis (#61)
* [DDW-1014] Trigger appropriate query based on input analysis

* [DDW-1014] Trigger appropriate query based on input analysis - querying input on analysis and adding Loading Spinner component

* [DDW-1014] Trigger correct query based on search input

The problem was that we have no easy way to
distinguish between transaction and block IDs
and for number inputs we have to ask the user
what he meant at least when there is an epoch
AND block with that number.

This commit solves the tx vs. block by fetching
both and checking the results + redirecting to
the corresponding search result page. The epoch
vs. block number problem was solved in the same
way but not perfectly: if you enter a number where
an epoch AND block exists it currently prefers the
epoch although we need to ask the user in the end.

* [DDW-1014] Trigger appropriate query based on input analysis - small fix

* Remove Header from Storybook as it's incompatible due to application logic coupling

* [DDW-1014] Fix storybook and search perf issues

1. The problem with storybook was that the feature
context was created within the index file where
also the api is imported, which resulted in the
graphql files being imported too, which we haven’t
setup in storybook (and shouldn’t). I separated
out the context and feature hooks into 
[feature]/context.ts files

2. The search performance issues where caused by
(1) not having global state that survives page
redirects. This was solved by moving the global
features to the special _app.tsx (the root for
next.js apps) which is not unmounted on redirects.
(2) the search store now checks if the requested
results are already loaded instead of blindly 
executing the search again.