Fixes #86 - Keep latest epochs up-to-date
The problem was that the latest epochs only have
been updated when a new epoch was available which
is not „realtime“ since the latest block date,
transactions and output are potentially changing
with each produced block.

The solution was to re-fetch the latest epochs
(a possible future optimization could be to only
re-fetch the current epoch) whenever the network
block height changes.

Additionally (because the data from the epochs / 
blocks endpoint are not 100% aligned) a reaction 
was added that overrides the lastBlockTime of the 
current epoch with the data from the latest 
fetched block, so that the obvious info is kept 
in sync in the UI.