[issue-169] Improve navigation feature
Motivation: The navigation feature is coupled to other features by specific actions that navigated to hardcoded pages. This meant adding new actions and store logic for any new page (which is ugly and takes more effort than necessary)

Solution: Each feature only knows about its own pages uses the generic navigation feature to interact with the underlying routing framework. This also improves portability and testability of the components because they are not directly tied to NextRouter anymore.

Extra benefit: the route path and query are now stored in the navigation store and can be observed like any other mobx observable.